+++ src/internal/transform/differentiation.zig
@@ -34,7 +34,7 @@ const Context = struct {
         }
 
         const result = switch (self.expression.node(id)) {
-            .integer, .rational, .float => self.builder.integer(0),
+            .integer, .rational, .float, .constant => self.builder.integer(0),
             .symbol => |name| self.builder.integer(
                 if (std.mem.eql(u8, name, self.variable)) 1 else 0,
             ),
@@ -163,6 +163,7 @@ const Context = struct {
             .integer => |value| self.builder.integer(value),
             .rational => |value| self.builder.rational(value),
             .float => |value| self.builder.float(value),
+            .constant => |value| self.builder.constant(value),
             .symbol => |name| self.builder.symbol(name),
             .add => |binary| self.builder.add(
                 self.clone(binary.left),